Player and Camera
Other than the score and Love Meter being set up ready to go for
implementation, the player movement (running, jumping), and camera
following the player is ready to go. Just need to generate platforms and
objects to collect/ not to collect, and sync that with the meter and the
score,, and have a highscore saved.