Speeding up and better ground detection
Added in a multiplier to be multiplied by the speed, and a target location so that after [x] amount of meters the speed seamlessly gets faster. Before, the player was able to lean against the side of the platform and jump, but now the bottom of the character needs to be touching the ground for the player to jump.