gameplay improvements
when the player hits the heart, and soon rose, the love metre and its text are updated. If the player collects a lot of items to increase the love metre it will never go over 100%. Next step is to create the pickup that decreases the  love metre and its text, and make both types of pickups spawn randomly.