GAME IS DONE
After a few weeks, and a couple of months, the game is finally finished. There was sound added in, then mostly touch ups and fine tuning took most of the time. It feels good enough to be released and played by people. Im happy I was able to hit my deadline, before Valentines Day.